refactor(omegaform): extract legacy _tag warning
