+++ packages/vue-components/src/components/OmegaForm/OmegaFormStuff.ts
@@ -20,20 +20,7 @@ import type {
   StringFieldMeta,
   UnknownFieldMeta
 } from "./meta/types"
-
-const legacyTagWarningEmittedFor = new Set<string>()
-type GlobalThisWithOptionalProcess = typeof globalThis & {
-  process?: {
-    env?: {
-      NODE_ENV?: string
-    }
-  }
-}
-
-const isDevelopmentEnvironment = () => {
-  const process = (globalThis as GlobalThisWithOptionalProcess).process
-  return process?.env?.NODE_ENV !== "production"
-}
+import { warnLegacyTag } from "./meta/legacyWarning"
 
 export type FieldPath<T> = unknown extends T ? string
   // technically we cannot have primitive at the root
@@ -904,18 +891,8 @@ const metadataFromAst = <From, To>(
             tagValue = resolvedTagType.literal as string
             discriminatorValues.push(tagValue)
             // Warn if the tag was wrapped in a single-element Union (legacy pattern)
-            if (
-              tagProp
-              && S.AST.isUnion(tagProp.type)
-              && isDevelopmentEnvironment()
-              && tagValue != null
-              && !legacyTagWarningEmittedFor.has(tagValue)
-            ) {
-              legacyTagWarningEmittedFor.add(tagValue)
-              console.warn(
-                `[OmegaForm] Union member with _tag "${tagValue}" uses S.Struct({ _tag: S.Literal("${tagValue}"), ... }). `
-                  + `Please migrate to S.TaggedStruct("${tagValue}", { ... }) for cleaner AST handling.`
-              )
+            if (tagProp && S.AST.isUnion(tagProp.type) && tagValue != null) {
+              warnLegacyTag(tagValue)
             }
           }
